+++ tests/src/utils.rs
@@ -24,6 +24,7 @@ use kube::{
     api::{DeleteParams, PostParams},
     Api, Client, Config, Error, Resource,
 };
+use kube_service::KubeService;
 pub use process::TestProcess;
 use rand::distr::{Alphanumeric, SampleString};
 use reqwest::{RequestBuilder, StatusCode};
@@ -41,6 +42,7 @@ use tokio::{
 
 pub(crate) mod application;
 pub(crate) mod ipv6;
+pub(crate) mod kube_service;
 pub(crate) mod port_forwarder;
 pub mod process;
 pub(crate) mod resource_guard;
@@ -210,61 +212,6 @@ pub async fn kube_client() -> Client {
     let mut config = Config::infer().await.unwrap();
     config.accept_invalid_certs = true;
     Client::try_from(config).unwrap()
-}
-
-/// A service deployed to the kubernetes cluster.
-///
-/// Service is meant as in "Microservice", not as in the Kubernetes resource called Service.
-/// This includes a Deployment resource, a Service resource and optionally a Namespace.
-pub struct KubeService {
-    pub name: String,
-    pub namespace: String,
-    pub service: Service,
-    pub deployment: Deployment,
-    guards: Vec<ResourceGuard>,
-    namespace_guard: Option<ResourceGuard>,
-    pub pod_name: String,
-}
-
-impl KubeService {
-    pub fn deployment_target(&self) -> String {
-        format!("deployment/{}", self.name)
-    }
-
-    pub fn pod_container_target(&self) -> String {
-        format!("pod/{}/container/{CONTAINER_NAME}", self.pod_name)
-    }
-}
-
-impl Drop for KubeService {
-    fn drop(&mut self) {
-        let mut deleters = self
-            .guards
-            .iter_mut()
-            .map(ResourceGuard::take_deleter)
-            .collect::<Vec<_>>();
-
-        deleters.push(
-            self.namespace_guard
-                .as_mut()
-                .and_then(ResourceGuard::take_deleter),
-        );
-
-        let deleters = deleters.into_iter().flatten().collect::<Vec<_>>();
-
-        if deleters.is_empty() {
-            return;
-        }
-
-        let _ = std::thread::spawn(move || {
-            tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()
-                .expect("failed to create tokio runtime")
-                .block_on(futures::future::join_all(deleters));
-        })
-        .join();
-    }
 }
 
 fn deployment_from_json(name: &str, image: &str, env: Value) -> Deployment {
